Add Creator class to hold methods responsible for loading StatTracker instance methods with data when they are instantiated.
